Add settings, improve accuracy, implement DTSpeed
Added a settings activity that allows the setting of two options. The
first turns on debug logging. The second adjusts how many decimal places
the user would like to see.
Began the process of moving from Double to BigDecimal for improved
accuracy. Still needs to be finished.
Began implementing the Data Transfer Speed conversions. On hold until
the move to BigDecimal is finished.