API 22, Support lib 22.1.1, Improved Nav Drawer
Moved to Android 5.1 (API 22). Of course, when the functionality of the
app is complete, the compatibility will be extended back to 4.0 (and
maybe even further).
Moved to the new support library, 22.1.1. Because of this,
ActionBarActivity is now deprecated, and it is now replaced with its
replacement, AppCompatActivity.
Moved to the native card implementation, rather than the custom one I
was using.
Improved the theme a bit, and the nav drawer now complies with material
guidelines (drawer drawn on top of everything, with a transparent
rectangle at the top for the status bar).
Added a base fragment for the other fragments to inherit from. This
allows a reduction in the amount of repeated code. The base fragment
class is expecting the root view of the layout given to it to be a
ScrollView. If something else is use, the app will crash.
Implemented a way for the toolbar to be hidden when scrolling, so that
there is more room on the screen. This is all taken care of by the base
fragment class.
Redid the debug option internally, so that it is better handled.
Reformatted/Refactored the code in various places for better and more
consistent formatting.

TODO:
- Fix a graphical bug that occurs upon returning to the main activity
from the settings
- Finish the internal state tracking when transitioning to the settings
activity and back
- Move the settings button from the overflow menu into the nav drawer
- Clean up the theme for the nav drawer
- Add in the other conversion types.