Updated Settings, Bug Fixes
Changed the settings activity to use a PreferenceFragment, rather than
completely rolling a custom layout. Made the corresponding changes to
the way internal settings are kept.
The graphical bug introduced in the last commit fixed itself.
Internal state tracking is finished and should work fine.
Some more reformatting was done.
Changed the base theme to be based on a light theme.

TODO:
- Move the settings button from overflow menu into nav drawer
- Clean up the theme for the nav drawer
- Add in the other conversion types.