More Material, Settings Icon
Instead of moving the settings item to the nav drawer, it was moved out
of the overflow menu, and given an icon.
Removed the manually drawn hamburger icon, and changed things so that
the toolbar draws its own, internal icon.
More material stuff. Toolbar title text is now in the proper location.
The nav drawer was cleaned up, and icons were added, along with things
complying with material guidelines. The settings icon ripple needs to be
tweaked.
The theme was tweaked to give it a more logical hierarchy.
Begun adding comments.

TODO:
- Tweak settings icon ripple.
- Properly test lifecycles, and make any necessary changes/tweaks.
- Look into boosting performance and reducing graphical hitches.
- Add in the other conversion types.
- Add the rest of the comments.