Use Getters/Setters, Decimal Setting Apply ASAP
In the base classes (activities and fragments), made all variables
private, and implemented getters and setters for the ones that should be
able to be modified by any activity or fragment that inherits from them.
Begun implementing a feature where, if the user had done a conversion
(meaning that there was text in the fields), and then switched to
settings and changed the decimal place count option, when they switched
back, the conversion updates itself to use the new setting. Implemented
and tested on the temperature fragment, and implemented, but not tested,
on the data transfer speed fragment. Need to test the latter fragment,
and finish the implementation and testing for the other fragments.
Begun implementation of classes that will allow the use of JUnit and
Instrumentation testing.
Now, when a fragment is created, the Activity is stored in a private
variable in the base fragment during onAttach(), and can be retrieved
using the getCurrentActivity() method. Should made it easier to handle
things that need a context.
In the Utils class, isDebugEnabled() will now return false if the
context provided is null.

TODO:
- Finish implementation and testing of the decimal place instant apply
feature
- Add in other conversions
- Finish testing framework
- TEST
- Add comments
- Release